pyproject.toml
[tool.poetry]
name = "stock"
version = "0.1.0"
description = ""
authors = ["kso12 <wookford@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
